Add fish 39 and 40 to anglerquestfish.py
The AnglerFishQuest tag can take a new value (40).

Currently, lihazahrd.World() crashes when trying to load a world with this value.

I verified ingame that a value of 40 corresponds to the Scorpio Fish.
This commit adds a value to the AnglerQuestFish enum for this new option.
It also adds a label for fish 39 (the Scarab Fish), which was previously unknown.

For backwards compatibility, I left the old name (UNKNOWN1) as-is.